Fix ftbxmodcompat crash by restoring TeamData compatibility
Restored `TeamData` class and made `IslandData` extend it to satisfy `ftbxmodcompat`'s field type expectations. Updated `ClientQuestFile` to expose `selfTeamData` as `TeamData` while using `selfIslandData` for internal logic. This resolves `java.lang.NoSuchFieldError: selfTeamData` while maintaining the custom IslandData functionality.
