+++ FTB-Quests-1.20.1-main/common/src/main/java/dev/ftb/mods/ftbquests/client/FTBQuestsClient.java
@@ -90,7 +90,7 @@ public class FTBQuestsClient {
 	}
 
 	public static IslandData getClientPlayerData() {
-		return ClientQuestFile.INSTANCE.selfTeamData;
+		return ClientQuestFile.INSTANCE.selfIslandData;
 	}
 
 	public static BaseQuestFile createClientQuestFile() {
