+++ FTB-Quests-1.20.1-main/common/src/main/java/dev/ftb/mods/ftbquests/client/gui/SelectQuestObjectScreen.java
@@ -70,7 +70,7 @@ public class SelectQuestObjectScreen<T extends QuestObjectBase> extends Abstract
 		ClientQuestFile file = ClientQuestFile.INSTANCE;
 		for (QuestObjectBase objectBase : file.getAllObjects()) {
 			if (config.predicate.test(objectBase) &&
-					(file.canEdit() || (!(objectBase instanceof QuestObject qo) || qo.isSearchable(file.selfTeamData)))) {
+					(file.canEdit() || (!(objectBase instanceof QuestObject qo) || qo.isSearchable(file.selfIslandData)))) {
 				list.add((T) objectBase);
 			}
 		}
